Fix rebuild on commit
