+++ crates/git_rev/README.md
@@ -1,3 +1,3 @@
 # Ambient git revision
 
-Tiny helper crate to get the current git revision.
+Tiny helper crate that holds the current git revision. The revision must be set in the final app to be available, use [git_rev_init](../git_rev_init/) for that.
