+++ crates/git_rev_init/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ambient_git_rev"
+name = "ambient_git_rev_init"
 version = { workspace = true }
 rust-version = { workspace = true }
 edition = "2021"
@@ -10,4 +10,4 @@ repository = "https://github.com/AmbientRun/Ambient"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-once_cell.workspace = true
+ambient_git_rev = { path = "../git_rev"}
