+++ crates/git_rev_init/README.md
@@ -1,0 +1,3 @@
+# Ambient git revision initializer
+
+Tiny helper crate to get the current git revision and set it in the holder provided by [git_rev](../git_rev/).
